add privacy policy link and cde optimization

+++ uicomponents/src/main/java/suncor/com/android/uicomponents/dropdown/ChildViewListener.java
@@ -1,5 +1,5 @@
 package suncor.com.android.uicomponents.dropdown;
 
 public interface ChildViewListener {
-        void onSelectFuelUpLimit(int value);
+        void onSelectFuelUpLimit(double value);
 }
